+++ extensions/manotori/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Manotori Changelog
 
+## [] - {PR_MERGE_DATE}
+
+- Better Error Handling
+- View details of Domain Event
+
 ## [Initial Version] - 2024-10-30
 
 - View Domains in Account